Add the triggers file and trigger data parsing tool
This was not tested very heavily, no idea if it works for all the
special trigger stuff.

The generator is debatable. For one, it's probably possibly to lazily
parse the parameters. Aside from that, it would be better if this sort
of stuff was part of something more standard that other tools could
benefit from as well, such as a master generator that extracts useful
data from a patch and puts them in a json format or something, and
then this tool uses that.