Add a "best fit" objects file; make objects files share a source file
The "best fit" strategy SHOULD always work as it should be impossible
that ability, upgrade or doodad files parse with the non-variation
format without errors and vice versa.

It would be better if the modification terminator would be consistent.
The modification terminator is supposed to be the object's that's being
modified ID (see ObjectModificationParentIdValidator), but it's
variously just 0. If it were consistent, it could be used to select the
best fit on a modification basis, and there wouldn't be a need for two
different file structs. Because it can be an ambiguous 0, however,
parsing is inconsistent. So here we have the "best fit" struct.