jass: Improve(?) tree construction
This is still a mess and inconsistent. I tried to make containers out
of the right things but it's very difficult to get right. So I'm now
giving up and leaving it as it is. Tools will just have to work around
this inconsistent/awkward structure as I've left it here. Until I
figure out a solid pattern to follow here anyway.