jass: Move grammar back to own variable
It should be clearer what the "struct" here is, which is the grammar,
so move it to the top where it's immediately seen. I'm still iffy on
the code below it, since the idea with this repo is to be definitions
only. Still, it makes more sense than the previous alternative.