jass: Add completely custom mappings for symbols
For better performance and memory use. Ideally this would've shortened
the lines of code here, but there were many things I didn't bother
splitting into functions since performance is the name of the game
here. A lot of this code is pretty dumb as a result, but shouldn't
need to be touched too much.