observer: Just accept utf8 strings for (flipped) byte strings
Will probably do this for the common ByteStringAdapter as well. Just
means when working with building you input "" instead of b"". As for
parsing, I think the name of the struct implies it's going to be a
string, and everyone I know is used to treating these "integer literal"
ids as basically strings.