Add storage formats (CASC, MPQ)
These originally belonged to the war3archiver repo:
https://github.com/warlockbrawl/war3archiver

Their commit history can be found there.

These are bindings to StormLib and CascLib. The DLLs are not yet part
of the repo so you need to build them yourself for now. Their addition
to this repo makes war3structs a comprehensive package for dealing with
all Warcraft III formats, whether maps or the actual installation.